requirements.txt
cx-Oracle==7.3.0
Django==2.2.7
mysql-connector==2.2.9
mysql-connector-python==8.0.18
mysqlclient==1.4.5
psycopg2-binary==2.8.3
PyExecJS==1.5.1
requests==2.18.4
